all code review nitpicks fixed.
