+++ s/runtests
@@ -20,6 +20,6 @@ $ROOT/$BUILD/eressea/eressea -v0 scripts/run-tests.lua
 $ROOT/$BUILD/eressea/eressea -v0 scripts/run-tests-e2.lua
 $ROOT/$BUILD/eressea/eressea -v0 scripts/run-tests-e3.lua
 $ROOT/$BUILD/eressea/eressea -v0 scripts/run-tests-e4.lua
-rm -rf data reports orders.txt score score.alliances
+rm -rf data reports orders.txt score score.alliances datum turn
 
 cd $OLDWPD
