+++ src/CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 2.8)
+cmake_minimum_required(VERSION 2.6)
 project (server C)
 
 include_directories (${CMAKE_CURRENT_SOURCE_DIR})
