+++ src/kernel/save.h
@@ -79,7 +79,6 @@ extern "C" {
 
     int write_game(struct gamedata *data);
     int read_game(struct gamedata *data);
-    int version_no(const char *str);
 
     /* test-only functions that give access to internal implementation details (BAD) */
     void _test_write_password(struct gamedata *data, const struct faction *f);
