+++ src/kernel/version.h
@@ -1,13 +1,17 @@
+#pragma once
 /* 
  +-------------------+  
  |                   |  Enno Rehling <enno@eressea.de>
  | Eressea PBEM host |  Katja Zedel <katze@felidae.kn-bremen.de>
- | (c) 1998 - 2007   |  Christian Schlittchen <corwin@amber.kn-bremen.de>
- |                   |
+ | (c) 1998 - 2016   |  Christian Schlittchen <corwin@amber.kn-bremen.de>
+ |                   |  https://github.com/eressea/server
  +-------------------+
-
- This program may not be used, modified or distributed 
- without prior permission by the authors of Eressea.
  */
 
+#ifndef H_KERNEL_VERSION
+#define H_KERNEL_VERSION
+
 const char *eressea_version(void);
+int version_no(const char *str);
+
+#endif
